backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-generativeai
groq
python-dotenv
